% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenny.R
\name{Jenny}
\alias{Jenny}
\title{This function returns Jenny's number.}
\usage{
Jenny(...)
}
\arguments{
\item{...}{Arguments passed on are ignored.}
}
\description{
This function returns Jenny's number.
}
\examples{
\dontrun{

phone_number = Jenny()
}
}
\author{
Anthogonyst
}
